+++ zsdk/zia/__init__.py
@@ -10,6 +10,7 @@ from .admin_roles import admin_roles
 from .admin_users import admin_users
 from .admin_audit_logs import admin_audit_logs
 from .device_groups import device_groups, devices
+from .event_logs import event_logs
 from .user_management import departments, groups, users, auditors
 from .firewall_policies import (
     firewall_filtering_rules,
@@ -202,3 +203,7 @@ class zia:
     @property
     def devices(self) -> devices:
         return devices(session=self._session, base_url=self._base_url)
+
+    @property
+    def event_logs(self) -> event_logs:
+        return event_logs(session=self._session, base_url=self._base_url)